UpdateHelper added to help with ESI client and cache DB updates
